+++ samples/wpf/Demo.NonModalCustomDialog/App.xaml.cs
@@ -1,0 +1,19 @@
+﻿using System.Windows;
+using CommunityToolkit.Mvvm.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
+using MvvmDialogs;
+
+namespace Demo.NonModalCustomDialog
+{
+    public partial class App
+    {
+        protected override void OnStartup(StartupEventArgs e)
+        {
+            Ioc.Default.ConfigureServices(
+                new ServiceCollection()
+                    .AddSingleton<IDialogService, DialogService>()
+                    .AddTransient<MainWindowViewModel>()
+                    .BuildServiceProvider());
+        }
+    }
+}