+++ CHANGELOG.md
@@ -6,6 +6,12 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ## Unreleased
 
+### :skull: Removed
+
+- [BREAKING CHANGE] Support for .NET Framework 3.5
+- [BREAKING CHANGE] Support for .NET Framework 4.0 and .NET Framework 4.5 due to [deprecation as of January 12, 2016](https://docs.microsoft.com/en-us/lifecycle/products/microsoft-net-framework)
+- [BREAKING CHANGE] Support for .NET Framework 4.5.2 due to [deprecation as of April 26, 2022](https://docs.microsoft.com/en-us/lifecycle/products/microsoft-net-framework)
+
 ## 8.0.0 - 2020-11-19
 
 ### :zap: Added
@@ -62,7 +68,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
         - `SafeFileNames` - An array that contains one safe file name for each selected file
         - `ValidateNames` - A value indicating whether the dialog accepts only valid Win32 file names
 
-### :syringe: Changed
+### :dizzy: Changed
 
 - [BREAKING CHANGE] The default value of `SaveFileDialogSettings.CheckFileExists` has changed from `true` to `false`, aligning it with the default value of the native Windows `SaveFileDialog`